requirements.txt
flask
flask-cors
python-dotenv
pytesseract
Pillow
PyPDF2
google-generativeai